==> 002_working_package_02/001_community_data/002_combined/001_diatoms/002_r_scripts/10_d_redundancy.R <==
# -------------------------------------- #
### --- Analyze Indicator Matrices --- ###
### ------------ Diatoms ------------- ###
### ------------ Redundancy   -------- ###
# -------------------------------------- #

# 06.11.20
# GetReal
# Working Package 2 
# Diatoms

source(file.path(DIR$rs, "10_c_setup_ta_analysis.R"))
source(file.path(DIR$red, "f_01_redundant.R"))

ma_redundnat = redundant(dt_bty)
print(ma_redundnat)
x11()
corrplot::corrplot.mixed(ma_redundnat, lower = "shade", upper = "number",
                         is.corr = FALSE,  order = "FPC") 
rm(redundant)

print("#----------------------------------------#")
